refactor: centralize range rewriting in `src/range.rs`
Move the range rewrite logic from `src/exrp.rs` to `src/range.rs`. There are a
few different places where we rewrite ranges so centralizing the logic in a
single location is ideal.
